requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
numpy==1.16.4
openapi-codec==1.3.2
pandas==0.24.2
psycopg2-binary==2.8.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
